some experimental fixes

+++ habitat_corl/dt.py
@@ -329,7 +329,7 @@ class DecisionTransformer(nn.Module):
         batch_size, seq_len = states.shape[0], states.shape[1]
         # [batch_size, seq_len, emb_dim]
         time_emb = self.timestep_emb(time_steps)
-        state_emb = self.state_emb(states) + time_emb
+        state_emb = self.state_emb(states.float()) + time_emb
         act_emb = self.action_emb(actions) + time_emb
         returns_emb = self.return_emb(returns_to_go.unsqueeze(-1)) + time_emb
